api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "MDF Dictionary Extractor — Indigenous languages PDF to Machine-Readable Dictionary Format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.12",
    "pymupdf>=1.27.2.2",
    "httpx>=0.27.0",
    "tiktoken>=0.8.0",
    "langchain>=1.2.15",
    "langchain-core>=0.3.0",
    "langchain-openai>=1.1.14",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-mistralai>=0.2.0",
    "langgraph>=1.1.8",
    "openai>=2.32.0",
    # OCR fallback (triggered automatically when native extraction fails)
    "pytesseract>=0.3.13",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
